refactor : 전략 패턴 적용
