+++ src/main/java/racingcar/domain/strategy/NumberGenerator.java
@@ -1,0 +1,5 @@
+package racingcar.domain.strategy;
+
+public interface NumberGenerator {
+    int generate();
+}
