+++ src/main/java/racingcar/domain/strategy/RandomNumberGenerator.java
@@ -1,12 +1,12 @@
-package racingcar.domain;
+package racingcar.domain.strategy;
 
 import java.util.Random;
 
-public class RandomNumberGenerator {
+public class RandomNumberGenerator implements NumberGenerator{
     private static final Random random = new Random();
     private static final int DEFAULT_BOUND = 10;
 
-    public static int generate() {
+    public int generate() {
         return random.nextInt(DEFAULT_BOUND);
     }
